+++ R/board.R
@@ -212,7 +212,7 @@ board_deregister <- function(name, ...) {
 #' board_register_local("temp", cache = tempfile())
 #'
 #' # configure default board
-#' options(pind.board = "temp")
+#' options(pins.board = "temp")
 #'
 #' # retrieve default board
 #' board_default()