+++ R/board_rsconnect.R
@@ -152,17 +152,16 @@ board_pin_create.rsconnect <- function(board, path, name, metadata, ...) {
   }
 }
 
-board_pin_find.rsconnect <- function(board, text = NULL, name = NULL, all_content = FALSE, ...) {
+board_pin_find.rsconnect <- function(board,
+                                     text = NULL,
+                                     name = NULL,
+                                     all_content = FALSE,
+                                     extended = FALSE,
+                                     ...) {
   if (is.null(text)) text <- ""
+  if (!is.null(name)) text <- pin_content_name(name)
 
-  if (nchar(text) == 0 && is.null(name)) {
-    # it can be quite slow to list all content in RStudio Connect so we scope to the user content
-    account_id <- rsconnect_api_get(board, "/__api__/users/current/")$id
-    filter <- paste0("filter=account_id:", account_id, "&accountId:", account_id)
-  }
-  else {
-    filter <- paste0("search=", text)
-  }
+  filter <- paste0("search=", text)
 
   entries <- rsconnect_api_get(board, paste0("/__api__/applications/?", utils::URLencode(filter)))$applications
   if (!all_content) entries <- Filter(function(e) e$content_category == "pin", entries)
@@ -173,6 +172,9 @@ board_pin_find.rsconnect <- function(board, text = NULL, name = NULL, all_conten
     name_pattern <- if (grepl("/", name)) paste0("^", name, "$") else paste0(".*/", name, "$")
     entries <- Filter(function(e) grepl(name_pattern, e$name), entries)
   }
+
+  if (identical(extended, TRUE))
+    return(pin_entries_to_dataframe(entries))
 
   results <- pin_results_from_rows(entries)
 
@@ -185,7 +187,11 @@ board_pin_find.rsconnect <- function(board, text = NULL, name = NULL, all_conten
   null_or_value <- function(e, value) if (is.null(e)) value else e
   results$name <- as.character(results$name)
   results$type <- unname(sapply(results$description, function(e) null_or_value(board_metadata_from_text(e)$type, "files")))
-  results$metadata <- sapply(results$description, function(e) as.character(jsonlite::toJSON(board_metadata_from_text(e), auto_unbox = TRUE)))
+
+  if (!identical(extended, TRUE)) {
+    results$metadata <- sapply(results$description, function(e) as.character(jsonlite::toJSON(board_metadata_from_text(e), auto_unbox = TRUE)))
+  }
+
   results$description <- board_metadata_remove(results$description)
 
   if (length(entries) == 1) {