+++ R/pin.R
@@ -191,6 +191,7 @@ pin_find_empty <- function() {
 #' @param text The text to find in the pin description or name.
 #' @param board The board name used to find the pin.
 #' @param name The exact name of the pin to match when searching.
+#' @param extended Should additional board-specific colulmns be shown?
 #' @param ... Additional parameters.
 #'
 #' @details
@@ -234,10 +235,15 @@ pin_find_empty <- function() {
 #' }
 #'
 #' @export
-pin_find <- function(text = NULL, board = NULL, name = NULL, ...) {
+pin_find <- function(text = NULL,
+                     board = NULL,
+                     name = NULL,
+                     extended = FALSE,
+                     ...) {
   if (is.null(board) || nchar(board) == 0) board <- board_list()
   metadata <- identical(list(...)$metadata, TRUE)
   text <- pin_content_name(text)
+  if (is.null(text) && !is.null(name)) text <- name
 
   all_pins <- pin_find_empty()
 
@@ -245,15 +251,28 @@ pin_find <- function(text = NULL, board = NULL, name = NULL, ...) {
     board_object <- board_get(board_name)
 
     board_pins <- tryCatch(
-      board_pin_find(board = board_object, text, name = name, ...),
+      board_pin_find(board = board_object, text, name = name, extended = extended, ...),
       error = function(e) {
         warning("Error searching '", board_name, "' board: ", e$message)
         board_empty_results()
       })
 
-    board_pins$board <- rep(board_name, nrow(board_pins))
+    if (identical(extended, TRUE)) {
+      ext_df <- tryCatch(
+        paste("[", paste(board_pins$metadata, collapse = ","), "]") %>% jsonlite::fromJSON(),
+        error = function(e) NULL)
 
-    all_pins <- rbind(all_pins, board_pins)
+      if (is.data.frame(ext_df) && nrow(board_pins) == nrow(ext_df)) {
+        ext_df <- ext_df[, !names(ext_df) %in% colnames(board_pins)]
+        board_pins <- cbind(board_pins, ext_df)
+      }
+    }
+
+    if (nrow(board_pins) > 0) {
+      board_pins$board <- rep(board_name, nrow(board_pins))
+
+      all_pins <- pin_results_merge(all_pins, board_pins, identical(extended, TRUE))
+    }
   }
 
   if (!is.null(text)) {
@@ -296,4 +315,115 @@ pin_files <- function(name, board = NULL, absolute = TRUE, ...) {
   metadata <- jsonlite::fromJSON(as.list(entry)$metadata)
 
   dir(file.path(board_local_storage(board), metadata$path), recursive = TRUE, full.names = absolute)
+}
+
+#' Pin Info
+#'
+#' Retrieve information for a given pin.
+#'
+#' @param name The exact name of the pin to match when searching.
+#' @param board The board name used to find the pin.
+#' @param extended Should additional board-specific colulmns be shown?
+#' @param ... Additional parameters.
+#'
+#' @examples
+#' library(pins)
+#'
+#' # define local board
+#' board_register_local(cache = tempfile())
+#'
+#' # cache the mtcars dataset
+#' pin(mtcars)
+#'
+#' # print pin information
+#' pin_info("mtcars")
+#'
+#' @export
+pin_info <- function(name, board = NULL, extended = TRUE, ...) {
+  entry <- pin_find(name = name, board = board, metadata = TRUE)
+
+  if (nrow(entry) == 0) stop("Pin '", name, "' was not found.")
+  if (nrow(entry) > 1) stop("Pin '", name, "' was found in multiple boards: ", paste(entry$board, llapse = ","),  ".")
+
+  board <- entry$board
+
+  metadata <- list()
+  if (!is.null(entry$metadata) && nchar(entry$metadata) > 0) {
+    metadata <- jsonlite::fromJSON(entry$metadata)
+  }
+
+  entry <- as.list(entry)
+  entry_ext <- list()
+
+  if (extended) {
+    entry_df <- pin_find(name = name, board = board, extended = TRUE)
+    if (nrow(entry_df) == 1) {
+      entry_ext <- as.list(entry_df)
+    }
+
+    entry_ext <- Filter(function(e) !is.list(e) || length(e) != 1 || !is.list(e[[1]]) || length(e[[1]]) > 0, entry_df)
+  }
+
+  for (name in names(metadata)) {
+    if (nrow(entry_ext) == length(metadata[[name]])) {
+      entry_ext[[name]] <- metadata[[name]]
+    }
+  }
+  entry$metadata <- NULL
+
+  for (name in names(entry)) {
+    entry_ext[[name]] <- entry[[name]]
+  }
+
+  structure(entry_ext, class = "pin_info")
+}
+
+print_pin_info <- function(name, e, ident) {
+  # avoid empty lavels that are nested
+  if (is.list(e) && is.null(names(e)) && length(e) == 1) e <- e[[1]]
+
+  # one-row data frames are better displayed as lists
+  if (is.data.frame(e) && nrow(e) == 1) e <- as.list(e)
+
+  if (!is.list(e) && is.vector(e)) {
+    cat(crayon::silver(paste0("#", ident, "- ", name, ": ", paste(e, collapse = ", "), "\n")))
+  }
+  else if (is.data.frame(e)) {
+    cat(crayon::silver(paste0("#", ident, "- ", name, ": ")))
+    if (length(colnames(e)) > 0) cat(crayon::silver(paste0("(", colnames(e)[[1]], ") ")))
+    cat(crayon::silver(paste(e[,1], collapse = ", ")))
+    if (length(colnames(e)) > 1) cat(crayon::silver("..."))
+    cat(crayon::silver("\n"))
+  }
+  else if (is.list(e)) {
+    cat(crayon::silver(paste0("#", ident, "- ", name, ": \n")))
+    for (i in names(e)) {
+      print_pin_info(i, e[[i]], paste0(ident, "  "))
+    }
+  }
+  else {
+    cat(crayon::silver(paste0("#", ident, "- ", name, ": ", class(e)[[1]], "\n")))
+  }
+}
+
+#' @keywords internal
+#' @export
+print.pin_info <- function(x, ...) {
+  info <- x
+
+  cat(crayon::silver(paste0("# Source: ", info$board, "<", info$name, "> [", info$type, "]\n")))
+  if (nchar(info$description) > 0) cat(crayon::silver(paste0("# Description: ", info$description, "\n")))
+
+  info$board <- info$name <- info$type <- info$description <- NULL
+
+  is_first <- TRUE
+  for (name in names(info)) {
+    e <- info[[name]]
+    if (identical(is.na(e), FALSE) && identical(is.null(e), FALSE) && !(is.character(e) && nchar(e) == 0)) {
+      if (is_first) cat(crayon::silver(paste0("# Extended:", "\n")))
+      is_first <- FALSE
+
+      print_pin_info(name, e, "   ")
+    }
+  }
 }